+++ src/core/brsTypes/interfaces/Draw2D.ts
